refactor(edge): extract quic worker lifecycle and shard wiring
