+++ crates/edge/Cargo.toml
@@ -35,6 +35,7 @@ rustls.workspace = true
 rustls-pki-types.workspace = true
 subtle.workspace = true
 x509-parser = "0.16"
+core_affinity.workspace = true
 
 [[test]]
 name = "regression"
